Do not generate entire taskgraph.  Instead just generate the slice that a particular processor cares about.  This creates a nice chicken/egg problem that we will have to worry about in a dynamic load-balancing world, but for now we will just focus on getting to that point...

+++ src/Packages/Uintah/CCA/Components/Schedulers/MemoryLog.cc
@@ -10,7 +10,7 @@ namespace Uintah {
 		 const std::string& label, const std::string& name,
 		 const std::string& type, const Patch* patch,
 		 int material, const std::string& nelems,
-		 int size, void* ptr, int dwid)
+		 unsigned long size, void* ptr, int dwid)
   {
     out << label;
     if(dwid != -1)
